Fix postinstall.sh for upgrading
If the plugin is already installed, this script failed on
`mkdir /etc/newrelic`. Change is to check if the dir exists before
attempting to create it.

Also, always copy the (potentially new) config into place. Additionally,
if there is no existing config present, copy the new config into place
for use by the plugin.